+++ service/src/main/java/uk/gov/mca/beacons/api/duplicates/domain/DuplicatesSummary.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class DuplicatesSummary {
 
   public String hexId;
-  public Long numberOfBeacons;
+  public Integer numberOfBeacons;
 }
